Simple pygame implementation in which “Goku” is our hero and he has to safeguard the dragonballs while the enemy is bombarding the dragon balls.
Simple pygame implementation in which “Goku” is our hero and he has to
safeguard the dragonballs while the enemy is bombarding the dragon balls. Goku uses his power
to neutralize enemy attacks. There is a timer and a accuracy meter which tells the time until which
the dragonballs were safe under goku’s protection and how accurately have goku used his power
respectively.